refactor(core): consolidate UserActionType into Core/Models
Move UserActionType from Data.Models and Telegram.Models into the
single canonical location Core/Models, eliminating the duplication
that was a latent drift hazard. The Data project cannot reference Core,
so UserActionRecordDto.ActionType is widened to int (EF HasConversion
is dropped as redundant) and all DTO-layer comparisons use (int)enum
casts. Application-layer code and UI models continue to use the typed enum.

+++ TelegramGroupsAdmin.IntegrationTests/Telegram/Services/WelcomeFlowBypassIntegrationTests.cs
@@ -279,7 +279,7 @@ public class WelcomeFlowBypassIntegrationTests
         var bypassRows = await context.UserActions
             .AsNoTracking()
             .Where(a => a.UserId == PreBannedUserTelegramId
-                        && a.ActionType == UserActionType.WelcomeBypass)
+                        && a.ActionType == (int)UserActionType.WelcomeBypass)
             .ToListAsync();
         Assert.That(bypassRows, Is.Empty,
             "Pre-banned user must not produce a bypass audit row — pre-ban short-circuit wins.");
@@ -350,7 +350,7 @@ public class WelcomeFlowBypassIntegrationTests
         var rows = await context.UserActions
             .AsNoTracking()
             .Where(a => a.UserId == telegramUserId
-                        && a.ActionType == UserActionType.WelcomeBypass)
+                        && a.ActionType == (int)UserActionType.WelcomeBypass)
             .ToListAsync();
 
         using (Assert.EnterMultipleScope())
